+++ packages/dnr-rulesets/src/cli.ts
@@ -1,10 +1,13 @@
 #!/usr/bin/env node
+import { Logger } from '@adguard/logger';
 import { program } from 'commander';
-import {
-    loadAssets,
-    patchManifest,
-} from './lib';
+
 import { version } from '../package.json';
+import {
+    AssetsLoader,
+    ManifestPatcher,
+    type PatchManifestOptions,
+} from './lib';
 
 program
     .name('dnr-rulesets CLI')
@@ -14,7 +17,17 @@ program
     .command('load')
     .description('Downloads rule sets for MV3 extension')
     .argument('[path-to-output]', 'rule sets download path')
-    .action(loadAssets);
+    .action(async (dest: string) => {
+        const logger = new Logger();
+        const loader = new AssetsLoader();
+
+        try {
+            await loader.load(dest);
+            logger.info(`assets was copied to ${dest}`);
+        } catch (e) {
+            logger.error(e);
+        }
+    });
 
 program
     .command('manifest')
@@ -23,6 +36,15 @@ program
     .argument('[path-to-filters]', 'filters src path')
     .option('-f, --force-update', 'force update rulesets with existing id', false)
     .option('-i, --ids <ids...>', 'filters ids to append', [])
-    .action(patchManifest);
+    .option('-e, --enable <ids...>', 'enable filters by default', [])
+    .option('-r, --ruleset-prefix <prefix>', 'prefix for filters ids', 'ruleset_')
+    .action((
+        manifestPath: string,
+        filtersPath: string,
+        options?: Partial<PatchManifestOptions>,
+    ) => {
+        const patcher = new ManifestPatcher();
+        patcher.patch(manifestPath, filtersPath, options);
+    });
 
 program.parse(process.argv);